requirements.txt
# --index-url https://download.pytorch.org/whl/cu124
# torch
# torchvision
# torchaudio
fastapi
uvicorn
jinja2
python-dotenv
opencv-python-headless
numpy
shapely
ultralytics
asyncio
scikit-image
glob2
filterpy
argparse
sqlalchemy
paho-mqtt
kafka-python
lz4